Esercizio completo MVC Classi

+++ MvcConsole/Controller.cs
@@ -23,25 +23,79 @@ class Controller{
                 SearchUserByName();
             } else if(input == "6")
             {
+                _db.CloseConnection();
+            }
+            else if(input == "7")
+            {
                 break;
             }
         }
     }
     private void DeleteUser(){
         Console.WriteLine("Enter user name u want to delete:");
-        var name = _view.GetInput();
-        _db.DeleteUser(name);
+        do
+        {
+            var name = _view.GetInput();
+            if (string.IsNullOrWhiteSpace(name))
+            {
+                Console.WriteLine("invalid input, try again");
+                continue;
+            }
+            else
+            {
+                _db.DeleteUser(name);
+                return;
+            }
+        }while (true);
     }
     private void UpdateUser(){
-        Console.WriteLine("Enter user name u want to update:");
-        var oldName = _view.GetInput();
-        var newName = _view.GetInput();
-        _db.UpdateUser(oldName,newName);
+        Console.WriteLine("Enter user name u want to update and both new name and surname:");
+        do
+        {
+            var active = true;
+            var oldName = _view.GetInput();
+            var newName = _view.GetInput();
+            var newSurname = _view.GetInput();
+            var choice = _view.GetInput();
+                if (choice == "0") active = false;
+                else if (choice == "1") active = true;
+                else Console.WriteLine("invalid input, try again");
+            if (string.IsNullOrWhiteSpace(oldName) || string.IsNullOrWhiteSpace(newName) || string.IsNullOrWhiteSpace(newSurname))
+            {
+                Console.WriteLine("invalid input, try again");
+                continue;
+            }
+            else
+            {
+                _db.UpdateUser(oldName,newName,newSurname,active);
+                return;
+            }
+        }while (true);
     }
     private void AddUser(){
-        Console.WriteLine("Enter user name:");
-        var name = _view.GetInput();
-        _db.AddUser(name);
+        Console.WriteLine("Enter user name and surname:");
+        do
+        {
+            var active = true;
+            var name = _view.GetInput();
+            var surname = _view.GetInput();
+            Console.WriteLine("Enter 0 for inactive state or 1 for active state");
+            var choice = _view.GetInput();
+                if (choice == "0") active = false;
+                else if (choice == "1") active = true;
+                else Console.WriteLine("invalid input, try again");
+
+            if (string.IsNullOrWhiteSpace(name) || string.IsNullOrWhiteSpace(surname))
+            {
+                Console.WriteLine("invalid input, try again");
+                continue;
+            }
+            else
+            {
+                _db.AddUser(name, surname, active);
+                return;
+            }
+        }while (true);
     }
     private void ShowUser(){
         var users = _db.GetUsers();
@@ -50,8 +104,20 @@ class Controller{
 
     private void SearchUserByName(){
         Console.WriteLine("Enter user name u want to search:");
-        var name = _view.GetInput();
-        var users = _db.SearchUserByName(name);
-        _view.ShowUsers(users);
+        do
+        {
+            var name = _view.GetInput();
+            if (string.IsNullOrWhiteSpace(name))
+            {
+                Console.WriteLine("invalid input, try again");
+                continue;
+            }
+            else
+            {
+                var users = _db.SearchUserByName(name);
+                _view.ShowUsers(users);
+                return;
+            }
+        }while (true);
     }
 }
